refactor composite logic
